+++ test/function/matrix/eye.test.js
@@ -14,7 +14,6 @@ describe('eye', function() {
     assert.deepEqual(math.eye(3,2), matrix([[1,0],[0,1],[0,0]]));
     assert.deepEqual(math.eye([3,2]), [[1,0],[0,1],[0,0]]);
     assert.deepEqual(math.eye(math.matrix([3,2])), matrix([[1,0],[0,1],[0,0]]));
-    assert.deepEqual(math.eye(math.matrix([[3],[2]])), matrix([[1,0],[0,1],[0,0]]));
     assert.deepEqual(math.eye(3,3), matrix([[1,0,0],[0,1,0],[0,0,1]]));
   });